refactor(app): split run.go (1082 lines) into construct + register phases (#138)
RunFree was a single 1082-line function mixing manager construction, service
registration, and error-path teardown. Split it into three single-responsibility
files plus a deps struct, with ZERO behavioral change:

- run.go (61 lines): thin orchestrator — buildAppDeps → registerServices →
  registerValues, with construction-cleanup on failure.
- builders.go (683): buildAppDeps — the construction phase (former Steps 0–10
  + HTTP router), verbatim, populating an appDeps struct.
- register.go (428): registerServices + registerValues — the registration
  phase (former services 1–16 + Values), verbatim, in the exact same order.
  Closures now read deps.* instead of local vars.
- deps.go (91): the appDeps struct holding all constructed managers.

Why not the issue's literal 7-file domain split (build_storage.go /
build_camera.go / etc.): the construction steps have cross-cutting injection
that a vertical domain split would force-restructure — healthMgr is built
then injected BACK into camMgr (SetHealthManager/SetRestarter/rediscovery),
mergeScheduler + periodicMergeManagers cross-inject and close over camMgr,
recordRollingMergeMgr/transcodeMgr/rollingMergeMgr are built before camMgr
(camMgr takes them as constructor args). Re-architecting those lateral
wirings is real risk for a refactor whose only goal is file size. The 2-phase
extract achieves the issue's actual goal (no single-file bottleneck, run.go
<300 lines) with no logic reordering.

Safety:
- No logic reordering — construction stays one function in original order;
  registration stays one function in original order. Only the file boundary
  moved.
- Order verified by TestRunFree_ServiceOrder (asserts the exact service-name
  sequence) + TestRunFree_SmokeStartStop + TestRunFree_StopJoinsBackgroundGoroutines.
- cmd/mibee-nvr/main.go untouched (RunFree signature unchanged).
- Values unchanged: same 6 RegisterValue calls in the same order.

All 4120 tests pass with -race; golangci-lint clean; gofmt clean.

Acceptance criteria (#138): run.go <300 lines ✅ (61), single-responsibility
files ✅, service start/stop order identical ✅, make test passes ✅,
main.go unchanged ✅.

Closes #138
